requirements.txt
flask~=1.1.2
nlpcube